+++ packages/core/monaco-editor/src/languages/yaml/context.ts
@@ -1,6 +1,6 @@
 import { parseDocument } from 'yaml'
 
-import type { YAMLModelContext } from '../types'
+import type { YAMLModelContext } from '../../types'
 
 export function parseIntoYAMLContext(value: string, altVersionId: number): Readonly<YAMLModelContext> {
   const document = parseDocument(value)