+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/side-panel/VaultField.vue
@@ -16,11 +16,11 @@
 
     <div class="vault-list">
       <template
-        v-for="entry in entries"
-        :key="entry.id"
+        v-for="([keyId, key]) in keys"
+        :key="keyId"
       >
         <div
-          v-if="entry.key"
+          v-if="key"
           class="vault-list-item"
         >
           <div class="vault-list-item-key">
@@ -29,7 +29,7 @@
               :size="16"
             />
             <div class="vault-list-item-key-text">
-              {{ entry.key }}
+              {{ key }}
             </div>
           </div>
           <KDropdown>
@@ -42,13 +42,13 @@
 
             <template #items>
               <KDropdownItem
-                @click="handleEdit(entry)"
+                @click="handleEdit(keyId)"
               >
                 {{ t('plugins.free-form.datakit.flow_editor.panel_segments.resources.vault.edit') }}
               </KDropdownItem>
               <KDropdownItem
                 danger
-                @click="handleRemove(entry)"
+                @click="handleRemove(keyId)"
               >
                 {{ t('plugins.free-form.datakit.flow_editor.panel_segments.resources.vault.delete') }}
               </KDropdownItem>
@@ -59,7 +59,7 @@
     </div>
 
     <VaultSecretPicker
-      :is-editing="!!selectedEntry && !creatingEntry"
+      :is-editing="!!selectedKeyId && !creatingKeyId"
       :secret-name="selectedEntry?.key"
       :secret-ref="selectedEntry?.value"
       :validate-name="validateName"
@@ -70,58 +70,72 @@
 </template>
 
 <script setup lang="ts">
-import { computed, ref } from 'vue'
+import { computed, ref, toRef } from 'vue'
 import composables from '../../../../../composables'
 import { AddIcon, MoreIcon, KeyIcon } from '@kong/icons'
 import VaultSecretPicker from '../VaultSecretPicker.vue'
-import type {
-  KeyValueFieldProps,
-  KeyValueFieldEmits,
-  KVEntry,
-} from '../../../shared/headless/useKeyValueField'
-import { useKeyValueField } from '../../../shared/headless/useKeyValueField'
 import {
   KUI_COLOR_TEXT_NEUTRAL,
   KUI_COLOR_TEXT,
 } from '@kong/design-tokens'
 import type { FieldName } from '../../types'
+import { useMapField, useField } from '../../../shared/composables'
+import type { KeyId } from '../../../shared/composables/key-id-map'
+import type { BaseFieldProps } from 'src/components/free-form/shared/types'
 
-interface Emits extends KeyValueFieldEmits {
+interface Emits {
   'update': [name: FieldName, value: string, oldName?: FieldName]
   'add': [name: FieldName, value: string]
   'remove': [name: FieldName]
 }
 
-type VaultKVEntry = KVEntry<FieldName, string>
-
-const props = defineProps<KeyValueFieldProps<FieldName>>()
+const props = defineProps<BaseFieldProps>()
 const emit = defineEmits<Emits>()
 
 const {
-  entries,
-  addEntry,
-  removeEntry,
-} = useKeyValueField<FieldName, string>(props, emit)
-const selectedEntry = ref<VaultKVEntry | null>(null)
-const creatingEntry = ref<VaultKVEntry | null>(null)
+  keys,
+  addKey,
+  removeKey,
+  updateKey,
+  getKeyName,
+} = useMapField<string, FieldName>(toRef(props, 'name'))
+
+const field = useField<Record<KeyId, string>>(toRef(props, 'name'))
+
+const selectedKeyId = ref<KeyId | null>(null)
+const creatingKeyId = ref<KeyId | null>(null)
+
+const selectedEntry = computed(() => {
+  if (!selectedKeyId.value) return null
+  const key = getKeyName(selectedKeyId.value)
+  const value = field.value ? field.value.value[selectedKeyId.value] : undefined
+  if (key === undefined || value === undefined) return null
+  return {
+    id: selectedKeyId.value,
+    key,
+    value,
+  }
+})
 
 const { i18n: { t } } = composables.useI18n()
 
 const vaultNames = computed(() => {
-  return new Set(entries.value.map((e) => e.key))
+  return new Set(keys.value.map(([, name]) => name))
 })
 
 const handleAdd = () => {
-  const newEntry = addEntry()
-  creatingEntry.value = newEntry
-  selectedEntry.value = newEntry
+  const fieldValue = field.value
+  if (!fieldValue) return
+  const kId = addKey()!
+  creatingKeyId.value = kId
+  selectedKeyId.value = kId
 }
 
 const closeVaultSecretPicker = () => {
-  selectedEntry.value = null
-  if (creatingEntry.value) {
-    removeEntry(creatingEntry.value.id)
-    creatingEntry.value = null
+  selectedKeyId.value = null
+  if (creatingKeyId.value) {
+    removeKey(creatingKeyId.value)
+    creatingKeyId.value = null
   }
 }
 
@@ -135,7 +149,15 @@ const handleVaultSecretSelected = (data: { secretRef: string, secretName: string
   selectedEntry.value!.value = data.secretRef
   selectedEntry.value!.key = data.secretName as FieldName
 
-  if (creatingEntry.value) {
+  if (hasValueChanged) {
+    field.value!.value[selectedEntry.value.id] = data.secretRef // Update freeform data
+  }
+
+  if (hasNameChanged) {
+    updateKey(selectedEntry.value.id, data.secretName as FieldName) // Update key map
+  }
+
+  if (creatingKeyId.value) {
     emit('add', data.secretName as FieldName, data.secretRef)
   } else if (hasChanged) {
     emit(
@@ -146,17 +168,18 @@ const handleVaultSecretSelected = (data: { secretRef: string, secretName: string
     )
   }
 
-  creatingEntry.value = null
+  creatingKeyId.value = null
   closeVaultSecretPicker()
 }
 
-const handleEdit = (entry: VaultKVEntry) => {
-  selectedEntry.value = entry
+const handleEdit = (kid: KeyId) => {
+  selectedKeyId.value = kid
 }
 
-const handleRemove = (entry: VaultKVEntry) => {
-  removeEntry(entry.id)
-  emit('remove', entry.key)
+const handleRemove = (kid: KeyId) => {
+  const keyName = getKeyName(kid)!
+  removeKey(kid)
+  emit('remove', keyName)
 }
 
 const validateName = (name: string, isEditing: boolean): string | undefined => {